+++ mangadex_library/lib/models/cover/Cover.dart
@@ -17,7 +17,7 @@ class Cover {
       limit = json['limit'] ?? 0;
       offset = json['offset'] ?? 0;
       total = json['total'] ?? 0;
-    } on Exception catch (e) {
+    } on Exception {
       throw MangadexServerException(json);
     }
   }