+++ mangadex_library/lib/models/user/user_followed_manga/manga_check.dart
@@ -8,7 +8,7 @@ class MangaCheck {
   MangaCheck.fromJson(Map<String, dynamic> json) {
     try {
       result = json['result'] ?? '';
-    } on Exception catch (e) {
+    } on Exception {
       throw MangadexServerException(json);
     }
   }